Fix memory leak in `RpcChannel`'s `wait_for_response`
Fix memory leak in the method `wait_for_response` of `RpcChannel`.
As documented in #15, when calling `asyncio.as_completed` it
actually creates a task to wait for the `self._close` event. In a
long lived `RpcChannel` that event never happens and thus that
future continues to run until the `close` method is called.
Since the task is created with each `RpcChannel`'s `call` method,
this could end up leaking objects such as `_asyncio.Task`,
`_asyncio.Future` and `coroutine`.

This change changes a the `wait_for_response` method to use
`asyncio.wait` with a `timeout` and `FIRST_COMPLETED` for the
`returned_when` argument such that we can obtain all pending futures
after the first completed and cancel them to avoid the leak. This
warrants that the `wait` for the `self._close` event future will be
cancelled and then no future/task/coroutine is leaked.

Also add some styling changes and fix some typos.

This fixes #15.